+++ www/src/components/utils/GoBack.js
@@ -2,15 +2,18 @@ import { A, Flex } from 'honorable'
 import { ArrowLeftIcon } from 'pluralsh-design-system'
 import { Link } from 'react-router-dom'
 
-export function GoBack({ text, link }) {
+export function GoBack({
+  text, link, as = Link, ...props
+}) {
   return (
     <Flex
       direction="row"
       paddingHorizontal="xxlarge"
       paddingVertical="large"
+      {...props}
     >
       <A
-        as={Link}
+        as={as}
         to={link}
         fontFamily="Monument Semi-Mono, monospace"
         fontWeight={500}
